Finish up Spot manipulation interface
